+++ core/events/transaction_result.go
@@ -28,6 +28,18 @@ type TransactionResult struct {
 	evt *eventspb.TransactionResult
 }
 
+func (tr *TransactionResult) PartyID() string {
+	return tr.evt.PartyId
+}
+
+func (tr *TransactionResult) Status() bool {
+	return tr.evt.Status
+}
+
+func (tr *TransactionResult) Hash() string {
+	return tr.evt.Hash
+}
+
 func NewTransactionResultEventSuccess(
 	ctx context.Context,
 	hash, party string,
@@ -184,6 +196,10 @@ func (t TransactionResult) IsParty(id string) bool {
 
 func (t TransactionResult) Proto() eventspb.TransactionResult {
 	return *t.evt
+}
+
+func (t TransactionResult) TransactionResult() TransactionResult {
+	return t
 }
 
 func (t TransactionResult) StreamMessage() *eventspb.BusEvent {